backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.2
django-cors-headers==4.3.1
pymongo[srv]==4.6.1
dnspython==2.6.1
python-dotenv==1.0.1
certifi==2024.2.2